Asks user to confirm action before deleting a goal.

+++ src/app/manufacturing-goals/manufacturing-goals.component.ts
@@ -11,6 +11,7 @@ import { auth } from '../auth.service';
 import { from } from 'rxjs';
 import { ActivityDetailsComponent } from '../activity-details/activity-details.component';
 import { EnableGoalsDialogComponent } from '../enable-goals-dialog/enable-goals-dialog.component';
+import { ConfirmActionDialogComponent } from '../confirm-action-dialog/confirm-action-dialog.component';
 
 export class ManufacturingGoal {
   activities: any[];
@@ -152,7 +153,12 @@ export class ManufacturingGoalsComponent implements OnInit {
   }
 
   deleteSelected(goal) {
-    this.deleteGoalConfirmed(goal.name);
+    const dialogConfig = new MatDialogConfig();
+    this.dialog.open(ConfirmActionDialogComponent, dialogConfig).afterClosed().subscribe(closeData => {
+        if (closeData && closeData['confirmed']) {
+          this.deleteGoalConfirmed(goal.name);
+        }
+      });
   }
 
   async deleteGoalConfirmed(name) {
